Move tag tests from Query.test.ts to TagsField.test.ts
They are still based on Query for now, so test TagsField indirectly.
